requirements.txt
py-cord==2.1.3
pypresence==4.2.1
dadjokes==1.3.2
nltk==3.7
pillow==8.4.0